Add Karaf build profile
Project local Karaf distros are handy for devs to test their work
however is unneeded by autorelease builds and should not be released
as part of the Simultanious Release. Add a profile that is active by
default so that default behaviour is unchanged however allows the
autorelease project to disable building this module.

Issue: RELENG-86
Change-Id: I7f57635b7e2baee8c3842550bfc14b109dc4ec8a
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>