Add network-policy model
This patch defines the k8s network-policy model. The model
is copied exactly from the upstream k8s model at:
https://github.com/kubernetes/kubernetes/blob/master/pkg/apis/
networking/types.go.
The k8s APIs include some definitions from other k8s apis
and those are in the core and meta-v1 yangs.

The aug.yang is the model for the extra metadata that
maps the label information to pods. This is a first pass
at how that would look. The idea is to augment the
selector grouping. This allows the base network-policy
model to remain clean and pure and any extra additions
are in the augmentation. That allows the different backends
to add whatever they need in the augmentation yangs.

JIRA: COE-34
Change-Id: I5704ce00edf3cd6e0912351b53c760ff5651817e
Signed-off-by: Sam Hague <shague@redhat.com>
(cherry picked from commit bd00fd74fec6311c33072ea57bbf7ba611fad7ea)